move mat3 to its own file
